refactor: movies.js로 영화 관련 API 분리 및 리팩토링 #29
